requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.6.15
charset-normalizer==2.1.1
Django==4.1
djangorestframework==3.13.1
idna==3.3
joblib==1.1.0
numpy==1.23.2
opencv-python==4.6.0.66
pandas==1.4.3
Pillow==9.2.0
pkg_resources==0.0.0
python-dateutil==2.8.2
pytz==2022.2.1
requests==2.28.1
scikit-learn==1.1.2
scipy==1.9.1
six==1.16.0
sqlparse==0.4.2
threadpoolctl==3.1.0
urllib3==1.26.12
